requirements.txt
flask
gevent
marshmallow
object-mapper
google-cloud-bigquery
python-dotenv
